refactor(ui): harden ForceGraph interactions
